move PipelineAppLogs to pipeline folder
